pyproject.toml
[tool.poetry]
name = "wemake-django-template"
version = "0.1.0"
description = "Bleeding edge django template focused on code quality and security."
authors = ["Melis <melis.zhoroev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "*"
django-split-settings = "*"
djangorestframework = "==3.8.2"
drf_yasg = "*"
drf-yasg = { version = "*", extras = ['validation'] }
djangorestframework-jwt = "==1.11.0"
python-telegram-bot = "==11.1.0"
celery = "*"
redis = "2.10.6"
django-celery = "3.2.2"
django-cache-memoize = "0.1.6"
flower = "*"
psycopg2 = "*"
gunicorn = "*"
python-decouple = "*"
dump-env = "*"
bcrypt = "*"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "*"
django-querycount = "*"
nplusone = "*"
xenon = "*"
radon = "*"
ipython = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
